# api/openapi.yaml
openapi: "3.0.3"
info:
  title: DataSpoke API
  version: "0.1.0"
  description: |
    DataSpoke is a sidecar extension to DataHub that provides user-group-specific features
    for Data Engineers (DE), Data Analysts (DA), and Data Governance personnel (DG).

    ## Three-Tier URI Structure

    ```
    /api/v1/spoke/common/…     Cross-cutting features shared by DE, DA, and DG
    /api/v1/spoke/dg/…         Data Governance features (requires "dg" group claim)
    /api/v1/hub/…              DataHub pass-through (any valid group)
    /health, /ready            System probes (no auth, no /api/v1 prefix)
    ```

    The `/spoke/de/` and `/spoke/da/` tiers are reserved but have no routes currently defined.
    All shared dataset operations (ingestion, validation, generation, search) live under
    `/spoke/common/` so that any team that owns a dataset can access them regardless of
    group membership.

    ## Access Control

    JWT Bearer tokens are required for all routes except `/health`, `/ready`, and `/auth/`.
    The `groups` claim in the JWT payload is an array of user-group identifiers.
    The middleware enforces that the claim matches the URI tier.

    | URI tier | Required group |
    |----------|----------------|
    | `/spoke/common/…` | any valid group (`de`, `da`, or `dg`) |
    | `/spoke/dg/…` | `dg` |
    | `/hub/…` | any valid group |
    | `/auth/…`, `/health`, `/ready` | public |
    | `/admin/…` | `admin` |

...

    ## WebSocket Channels

...

...

...

    ## Pagination

...

...
servers:
...
paths:
...
  /spoke/dg/metric:
    get:
      operationId: listMetrics
      summary: List all governance metrics
      description: Paginated list of metrics. Filterable by theme and status.
      tags: [dg]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sort"
        - name: theme
          in: query
          description: Filter by metric theme
          schema:
            type: string
        - name: status
          in: query
          description: Filter by metric active status
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Paginated list of metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: getMetric
      summary: Get metric summary
      description: Returns identity, theme, and active status.
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: getMetricAttr
      summary: Get metric attributes overview
      description: Returns theme, period, active status, and alarm enabled flag.
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric attributes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricAttr"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr/conf:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: getMetricConf
      summary: Get full metric definition
      description: Returns title, theme, measurement period, alarm setup, and active status.
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Full metric definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricConf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      operationId: putMetricConf
      summary: Create or replace metric definition
      tags: [dg]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricConfWrite"
      responses:
        "200":
          description: Metric definition replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricConf"
        "201":
          description: Metric definition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricConf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      operationId: patchMetricConf
      summary: Update metric definition fields
      tags: [dg]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricConfPatch"
      responses:
        "200":
          description: Metric definition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricConf"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteMetricConf
      summary: Remove metric definition
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Metric definition deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr/result:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: getMetricResult
      summary: Get measurement results for metric
      description: Returns numeric timeseries. Use `?from=…&to=…` to narrow the time range.
      tags: [dg]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sort_occurred_at"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
      responses:
        "200":
          description: Paginated metric measurement results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricResultList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/method/run:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    post:
      operationId: runMetric
      summary: Trigger a metric measurement run
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric measurement run triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunTriggered"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/method/activate:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    post:
      operationId: activateMetric
      summary: Activate metric (enable scheduled measurement)
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/method/deactivate:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    post:
      operationId: deactivateMetric
      summary: Deactivate metric
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/event:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: listMetricEvents
      summary: Get metric run events and alarm notices
      tags: [dg]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sort_occurred_at"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
      responses:
        "200":
          description: Paginated metric event log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ==========================================================================
  # DG — Metric Issues
  # ==========================================================================
  /spoke/dg/metric/{metric_id}/attr/issue:
    parameters:
      - $ref: "#/components/parameters/metric_id"
    get:
      operationId: listMetricIssues
      summary: List metric issues
      description: |
        Auto-detected metric issues with lifecycle tracking. Filterable by
        status, priority, issue_type, and assignee.
      tags: [dg]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          description: Filter by metric issue status
          schema:
            type: string
            enum: [open, in_progress, resolved, dismissed]
        - name: priority
          in: query
          description: Filter by metric issue priority
          schema:
            type: string
            enum: [critical, high, medium]
        - name: issue_type
          in: query
          description: Filter by metric issue type (e.g. missing_owner, no_description, stale)
          schema:
            type: string
        - name: assignee
          in: query
          description: Filter by assignee email
          schema:
            type: string
      responses:
        "200":
          description: Paginated metric issues list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricIssueList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr/issue/{metric_issue_id}:
    parameters:
      - $ref: "#/components/parameters/metric_id"
      - $ref: "#/components/parameters/metric_issue_id"
    get:
      operationId: getMetricIssue
      summary: Get metric issue detail
      description: |
        Returns full metric issue detail including type, priority, status, assignee,
        projected score impact, and estimated fix time.
      tags: [dg]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Metric issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricIssueDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      operationId: patchMetricIssue
      summary: Update metric issue fields
      description: |
        Update mutable metric issue fields: status (open, in_progress), assignee, due_date.
        To dismiss a metric issue use the dedicated method/dismiss endpoint instead.
      tags: [dg]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricIssuePatch"
      responses:
        "200":
          description: Metric issue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricIssueDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr/issue/{metric_issue_id}/method/dismiss:
    parameters:
      - $ref: "#/components/parameters/metric_id"
      - $ref: "#/components/parameters/metric_issue_id"
    post:
      operationId: dismissMetricIssue
      summary: Dismiss metric issue
      description: |
        Mark a metric issue as dismissed — acknowledged but will not fix. This is a
        deliberate business decision distinct from resolution (which is automatic).
      tags: [dg]
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Optional reason for dismissal
      responses:
        "200":
          description: Metric issue dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricIssueDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /spoke/dg/metric/{metric_id}/attr/issue/{metric_issue_id}/event:
    parameters:
      - $ref: "#/components/parameters/metric_id"
      - $ref: "#/components/parameters/metric_issue_id"
    get:
      operationId: listMetricIssueEvents
      summary: Get metric issue lifecycle events
      description: |
        Immutable log of metric issue lifecycle transitions: creation, status changes,
        assignment changes, and dismissal reasons.
      tags: [dg]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sort_occurred_at"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
      responses:
        "200":
          description: Paginated metric issue event log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ==========================================================================
  # DG — Overview
  # ==========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via `POST /auth/token`. Pass as
        `Authorization: Bearer <access_token>`.

  # --------------------------------------------------------------------------
  # Reusable Parameters
  # --------------------------------------------------------------------------
  parameters:
...
    metric_id:
      name: metric_id
      in: path
      required: true
      description: Governance metric identifier
      schema:
        type: string
...
    offset:
      name: offset
      in: query
      description: Pagination start index (0-based)
      schema:
        type: integer
        default: 0
        minimum: 0
    limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    sort:
      name: sort
      in: query
      description: |
        Sort field and direction as `{field}_{asc|desc}`, e.g. `created_at_desc`.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Malformed request, missing required fields, or invalid parameter values
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: INVALID_PARAMETER
            message: "The 'limit' field must be an integer between 1 and 100."
            trace_id: "550e8400-e29b-41d4-a716-446655440000"
    Unauthorized:
      description: Missing, expired, or malformed access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: UNAUTHORIZED
            message: "Access token is missing or has expired."
            trace_id: "550e8400-e29b-41d4-a716-446655440000"
    Forbidden:
      description: Valid token but insufficient group claim for this route
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: FORBIDDEN
            message: "Your account does not have access to this resource."
            trace_id: "550e8400-e29b-41d4-a716-446655440000"
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # --------------------------------------------------------------------------
  # Reusable Schemas
  # --------------------------------------------------------------------------
  schemas:
...
    MetricSummary:
      allOf:
        - $ref: "#/components/schemas/RespTimeMixin"
        - type: object
          required: [metric_id, title, theme, active]
          properties:
            metric_id:
              type: string
            title:
              type: string
            theme:
              type: string
              description: "Metric theme, e.g. documentation, ownership, freshness, errors"
            active:
              type: boolean
            created_at:
              type: string
              format: date-time
    MetricList:
      allOf:
        - $ref: "#/components/schemas/PaginationMeta"
        - type: object
          required: [metrics]
          properties:
            metrics:
              type: array
              items:
                $ref: "#/components/schemas/MetricSummary"
...
    MetricConfWrite:
      allOf:
        - $ref: "#/components/schemas/MetricConfBase"
...